Add new movie endpoint to search movies by id

+++ plex-service/src/main/java/com/cleverpine/plex/repository/future/MoviesRepository.java
@@ -3,5 +3,5 @@ package com.cleverpine.plex.repository.future;
 import com.cleverpine.plex.entity.future.MovieEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoviesRepository extends JpaRepository<MovieEntity, Integer> {
+public interface MoviesRepository extends JpaRepository<MovieEntity, Long> {
 }
